Reimplement ai widget as an extension
