+++ scripts/extensions/ai-widget/.gitignore
@@ -1,0 +1,1 @@
+dist
