+++ scripts/extensions/ai-widget/package.json
@@ -1,0 +1,16 @@
+{
+  "private": true,
+  "main": "dist/src/extension.js",
+  "scripts": {
+    "compile": "tsc -p src",
+    "compile-e2e": "echo 'No end-to-end tests defined'",
+    "watch": "tsc -watch -p src"
+  },
+  "devDependencies": {
+    "superdesk-code-style": "1.3.0",
+    "typescript": "^4.5.2"
+  },
+  "dependencies": {
+    "@superdesk/common": "^0.0.8"
+  }
+}
