+++ scripts/extensions/ai-widget/src/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "superdesk-code-style/tsconfig-strict",
+    "ts-node": {
+        "files": true // needed for mocha to recognize src/typings/refs.d.ts
+    },
+    "compilerOptions": {
+        "outDir": "./dist/src",
+        "declaration": true,
+        "esModuleInterop": true,
+    }
+}
