+++ scripts/extensions/ai-widget/src/typings/refs.d.ts
@@ -1,0 +1,1 @@
+/// <reference path='../../../../core/superdesk-api.d.ts' />
